# auth-service/docs/swagger.yaml
basePath: /
definitions:
  dto.AuthResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
host: localhost:8080
info:
  contact: {}
  title: Auth Service API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: invalid json / validation error
          schema:
            type: string
        "401":
          description: invalid credentials
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Login user
      tags:
      - auth
  /refresh-token:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: invalid json / validation error
          schema:
            type: string
        "401":
          description: invalid credentials
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Refresh access token
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: invalid json
          schema:
            type: string
        "409":
          description: email already exists
          schema:
            type: string
      summary: Register user
      tags:
      - auth
swagger: "2.0"